requirements.txt
Flask==3.0.2
python-pptx==0.6.21

google-generativeai==0.4.0